+++ Esp.Net.Tests/Esp.Net.Tests.csproj
@@ -50,7 +50,9 @@
     <Compile Include="..\AssemblyInfoCommon.cs">
       <Link>Properties\AssemblyInfoCommon.cs</Link>
     </Compile>
-    <Compile Include="Pipeline\PipelineTests.cs" />
+    <Compile Include="Concurrency\BeginAcyncTests.cs" />
+    <Compile Include="Concurrency\PipelineTests.cs" />
+    <Compile Include="Concurrency\TestSubject.cs" />
     <Compile Include="Reactive\ReactiveTests.cs" />
     <Compile Include="RouterTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
