+++ Esp.Net/Concurrency/AsyncResultsEvent.cs
@@ -1,0 +1,17 @@
+using System;
+
+namespace Esp.Net.Concurrency
+{
+    public class AsyncResultsEvent<TResult>
+    {
+        public AsyncResultsEvent(TResult result, Guid id)
+        {
+            Result = result;
+            Id = id;
+        }
+
+        public TResult Result { get; private set; }
+
+        public Guid Id { get; private set; }
+    }
+}
